Add files via upload
The 'Diamonds' dataset:

Released in 2007, the 'Diamonds' dataset is a well known, easily accessible dataset that contains the price, as well as 9 other attributes (such as carat, cut, clarity, etc.), of 53940 round-cut diamonds sourced from the 'Loose Diamonds Search Engine'. Since its release, it has become a popular 'practice' dataset for those seeking to hone their machine-learning and data-analyses skills. It is most often used to practice data cleaning and preprocessing, as well as regression modelling, with hundreds of "Predicting the Price of Diamonds using Regression" notebooks having been created and uploaded to sites such as Kaggle over the years. The original dataset, of which this is a copy, can be found at - https://www.kaggle.com/datasets/shivam2503/diamonds.